lxd/instance/qemu/qmp: Re-shuffle functions
